Main restaurant ready

+++ src/person/Employee.java
@@ -2,11 +2,11 @@ package person;
 
 public class Employee extends Person{
     private String job;
-    private Double salary;
+    private int salary;
     private Boolean state;
 
     // Constructor
-    public Employee(String fullName, String phone, String gender, Byte age, String job, Double salary, Boolean state){
+    public Employee(String fullName, String phone, String gender, Byte age, String job, int salary, Boolean state){
         super(fullName, phone, gender, age);
         this.job = job;
         this.salary = salary;
@@ -17,7 +17,7 @@ public class Employee extends Person{
     public String getJob(){
         return this.job;
     }
-    public Double getSalary(){
+    public int getSalary(){
         return this.salary;
     }
     public Boolean getState(){
@@ -27,7 +27,7 @@ public class Employee extends Person{
     public void setJob(String job){
         this.job = job;
     }
-    public void setSalary(Double salary){
+    public void setSalary(int salary){
         this.salary = salary;
     }
     public void setState(Boolean state){
